requirements.txt
fastapi==0.100.0
uvicorn==0.23.2
python-dotenv==1.0.1
openai==0.28.1
